+++ system/DataCaster/Cast/DatetimeCast.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace CodeIgniter\DataConverter\Cast;
+namespace CodeIgniter\DataCaster\Cast;
 
 use CodeIgniter\I18n\Time;
