+++ system/DataCaster/Exceptions/CastException.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace CodeIgniter\DataConverter\Exceptions;
+namespace CodeIgniter\DataCaster\Exceptions;
 
 use CodeIgniter\Entity\Exceptions\CastException as EntityCastException;
