--- system/DataConverter/Cast/ArrayCast.php
@@ -1,43 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * This file is part of CodeIgniter 4 framework.
- *
- * (c) CodeIgniter Foundation <admin@codeigniter.com>
- *
- * For the full copyright and license information, please view
- * the LICENSE file that was distributed with this source code.
- */
-
-namespace CodeIgniter\DataConverter\Cast;
-
-/**
- * Class ArrayCast
- *
- * (PHP) [array --> string] --> (DB driver) --> (DB column) string
- *       [      <-- string] <-- (DB driver) <-- (DB column) string
- *
- * @extends BaseCast<mixed[], string, mixed>
- */
-class ArrayCast extends BaseCast implements CastInterface
-{
-    public static function get(mixed $value, array $params = []): array
-    {
-        if (! is_string($value)) {
-            self::invalidTypeValueError($value);
-        }
-
-        if ((strpos($value, 'a:') === 0 || strpos($value, 's:') === 0)) {
-            $value = unserialize($value, ['allowed_classes' => false]);
-        }
-
-        return (array) $value;
-    }
-
-    public static function set(mixed $value, array $params = []): string
-    {
-        return serialize($value);
-    }
-}
